@@ -33,31 +33,31 @@ class SignInDemo extends StatefulWidget {
 }
 
 class SignInDemoState extends State<SignInDemo> {
-  GoogleSignInAccount _currentUser;
-  String _contactText;
+  GoogleSignInAccount? _currentUser;
+  String _contactText = '';
 
   @override
   void initState() {
     super.initState();
-    _googleSignIn.onCurrentUserChanged.listen((GoogleSignInAccount account) {
+    _googleSignIn.onCurrentUserChanged.listen((GoogleSignInAccount? account) {
       setState(() {
         _currentUser = account;
       });
       if (_currentUser != null) {
-        _handleGetContact();
+        _handleGetContact(_currentUser!);
       }
     });
     _googleSignIn.signInSilently();
   }
 
-  Future<void> _handleGetContact() async {
+  Future<void> _handleGetContact(GoogleSignInAccount user) async {
     setState(() {
       _contactText = "Loading contact info...";
     });
     final http.Response response = await http.get(
-      'https://people.googleapis.com/v1/people/me/connections'
-      '?requestMask.includeField=person.names',
-      headers: await _currentUser.authHeaders,
+      Uri.parse('https://people.googleapis.com/v1/people/me/connections'
+          '?requestMask.includeField=person.names'),
+      headers: await user.authHeaders,
     );
     if (response.statusCode != 200) {
       setState(() {
@@ -68,7 +68,7 @@ class SignInDemoState extends State<SignInDemo> {
       return;
     }
     final Map<String, dynamic> data = json.decode(response.body);
-    final String namedContact = _pickFirstNamedContact(data);
+    final String? namedContact = _pickFirstNamedContact(data);
     setState(() {
       if (namedContact != null) {
         _contactText = "I see you know $namedContact!";
@@ -78,14 +78,14 @@ class SignInDemoState extends State<SignInDemo> {
     });
   }
 
-  String _pickFirstNamedContact(Map<String, dynamic> data) {
-    final List<dynamic> connections = data['connections'];
-    final Map<String, dynamic> contact = connections?.firstWhere(
+  String? _pickFirstNamedContact(Map<String, dynamic> data) {
+    final List<dynamic>? connections = data['connections'];
+    final Map<String, dynamic>? contact = connections?.firstWhere(
       (dynamic contact) => contact['names'] != null,
       orElse: () => null,
     );
     if (contact != null) {
-      final Map<String, dynamic> name = contact['names'].firstWhere(
+      final Map<String, dynamic>? name = contact['names'].firstWhere(
         (dynamic name) => name['displayName'] != null,
         orElse: () => null,
       );
@@ -107,26 +107,27 @@ class SignInDemoState extends State<SignInDemo> {
   Future<void> _handleSignOut() => _googleSignIn.disconnect();
 
   Widget _buildBody() {
-    if (_currentUser != null) {
+    GoogleSignInAccount? user = _currentUser;
+    if (user != null) {
       return Column(
         mainAxisAlignment: MainAxisAlignment.spaceAround,
         children: <Widget>[
           ListTile(
             leading: GoogleUserCircleAvatar(
-              identity: _currentUser,
+              identity: user,
             ),
-            title: Text(_currentUser.displayName ?? ''),
-            subtitle: Text(_currentUser.email ?? ''),
+            title: Text(user.displayName ?? ''),
+            subtitle: Text(user.email),
           ),
           const Text("Signed in successfully."),
-          Text(_contactText ?? ''),
+          Text(_contactText),
           ElevatedButton(
             child: const Text('SIGN OUT'),
             onPressed: _handleSignOut,
           ),
           ElevatedButton(
             child: const Text('REFRESH'),
-            onPressed: _handleGetContact,
+            onPressed: () => _handleGetContact(user),
           ),
         ],
       );
